fix evenly

+++ include/ContainerPlus.hpp
@@ -38,11 +38,11 @@ namespace stdplus
 	template<typename T>
 	std::vector<T> evenly(const std::vector<T> vec, int dev)
 	{
-		int size = vec.size();
-		int step = size / dev;
+		size_t size = vec.size();
+		size_t step = size / dev;
 		std::vector<T> result;
 
-		for (int i = 0; i < size; i += step)
+		for (size_t i = 0; i < size; i += step)
 		{
 			result.push_back(vec.at(i));
 			if (result.size() == dev)
